api/requirements.txt
firebase_admin
flasgger
Flask==2.0.1
dnspython